requirements.txt
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Flask==2.3.3
google-ai-generativelanguage==0.4.0
google-api-core==2.24.2
google-auth==2.38.0
google-generativeai==0.3.1
googleapis-common-protos==1.69.2
grpcio==1.71.0
grpcio-status==1.62.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
proto-plus==1.26.1
protobuf==4.25.6
pyasn1==0.6.1
pyasn1_modules==0.4.2
python-dotenv==1.0.0
requests==2.32.3
rsa==4.9
tqdm==4.67.1
urllib3==2.3.0
Werkzeug==3.1.3
